ADM-973 [frontend] feat: extract ReportContent from Report Step
